Entity has item_type working
When an entity requests item_type we get the correct result. item_type_values is not working because it is a polymorphic relation. When values are stored type is set to Entity, but when they are retrieved they are searched for either "Item" or "Collection". Need to look into this.